consolidating test and udpate into one
